separate classes, added feedback, organized gameloop
